Fork inline link test for LinkV1
